+++ src/components/RuleBuilderDialog.jsx
@@ -1,0 +1,44 @@
+// src/components/RuleBuilderDialog.jsx
+import React from 'react';
+
+export default function RuleBuilderDialog({ open, title, description, onClose, children }) {
+  if (!open) return null;
+
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/40 p-4">
+      <div className="w-full max-w-4xl overflow-hidden rounded-lg border border-gray-200 bg-white shadow-xl dark:border-gray-800 dark:bg-gray-900">
+        <div className="flex items-start justify-between gap-4 border-b border-gray-200 px-4 py-4 dark:border-gray-800 sm:px-6">
+          <div className="space-y-1">
+            <h3 className="text-lg font-semibold text-gray-900 dark:text-gray-50">{title}</h3>
+            {description && (
+              <p className="text-sm text-gray-600 dark:text-gray-400">{description}</p>
+            )}
+          </div>
+          <button
+            type="button"
+            onClick={onClose}
+            className="inline-flex items-center rounded-md border border-gray-300 p-2 text-gray-500 transition hover:bg-gray-100 dark:border-gray-700 dark:text-gray-300 dark:hover:bg-gray-800"
+            aria-label="Close builder dialog"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="20"
+              height="20"
+              viewBox="0 0 24 24"
+              fill="none"
+              stroke="currentColor"
+              strokeWidth="2"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              className="size-4"
+            >
+              <line x1="18" x2="6" y1="6" y2="18" />
+              <line x1="6" x2="18" y1="6" y2="18" />
+            </svg>
+          </button>
+        </div>
+        <div className="max-h-[80vh] overflow-y-auto px-4 py-4 sm:px-6">{children}</div>
+      </div>
+    </div>
+  );
+}
